Day 18 part 2

+++ AdventOfCode23Day10/Program.cs
@@ -6,7 +6,7 @@ string input = Resources.Input1;
 Map map = new(input.Split(Environment.NewLine));
 
 int stepsToFurthestPartOfLoop = map.Loop.Count / 2;
-int enclosedArea = map.FindEnclosedArea();
+long enclosedArea = map.FindEnclosedArea();
 
 Console.WriteLine($"Steps to furthest part of loop: {stepsToFurthestPartOfLoop}");
 Console.WriteLine();
